backend/requirements.txt
flask
requests
psycopg2
Flask-Login
Werkzeug
Flask-CORS
python-dotenv